ioquake3 porting

+++ code/sys/con_log.c
@@ -60,7 +60,7 @@ CON_LogWrite
 */
 unsigned int CON_LogWrite( const char *in )
 {
-	unsigned int length = (unsigned int)strlen( in );
+	unsigned int length = strlen( in );
 	unsigned int firstChunk;
 	unsigned int secondChunk;
